tests/wpt/web-platform-tests/encoding-detection/pl-windows-1250-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>pl windows-1250 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "windows-1250", 'Expected windows-1250');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/pl-windows-1250-late.sub.html");
</script>
